Removed flow obfuscation/obfuscation-related exception handlers

+++ client/src/main/java/sign/signlink.java
@@ -411,7 +411,7 @@ public final class signlink implements Runnable {
 				urlreq = null;
 			}
 			try {
-				Thread.sleep((long) looprate);
+				Thread.sleep(looprate);
 			} catch (@Pc(187) Exception local187) {
 			}
 		}
